Move Auth error notice out of routes
